Fix type casting for size calculations and update member variable type for clarity

+++ src/CKFile.cpp
@@ -116,7 +116,7 @@ char *CKBufferParser::ReadString() {
 }
 
 int CKBufferParser::ReadInt() {
-    if (m_CursorPos + sizeof(int) > m_Size)
+    if (m_CursorPos + (int) sizeof(int) > m_Size)
         return 0;
 
     int val;
